% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indv_change}
\alias{indv_change}
\title{Repeated measures data over three years}
\format{A data frame with 234 observations and 5 variables:
\describe{
  \item{id}{subject identifier}
  \item{cycle}{cycle number}
  \item{cycle_time}{time since the start of the cycle, in weeks}
  \item{start_time}{time since the start of the study, in weeks}
  \item{response}{outcome measure}
}}
\usage{
indv_change
}
\description{
A dataset of longitudinal responses of 36 study participants over a three years span.
}
\keyword{datasets}

