% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mean_change}
\alias{mean_change}
\title{Data for a single time trend over three years}
\format{A data frame with 9 observations and 4 variables:
\describe{
  \item{cycle}{cycle number}
  \item{cycle_time}{time since the start of the cycle, in weeks}
  \item{start_time}{time since the start of the study, in weeks}
  \item{response}{outcome measure}
}}
\usage{
mean_change
}
\description{
A dataset of mean responses of study participants over a three years span.
}
\keyword{datasets}

